improved test script

+++ test.sh
@@ -1,9 +1,23 @@
-#!/bin/bash -uex
+#!/bin/bash -ue
 
-# cleanup previous box, create new box and start ssh login
-vagrant suspend
-vagrant destroy -f
-vagrant box remove -f funtoo-stage3
-vagrant box add --name funtoo-stage3 funtoo-x86_64-generic_64-stage3.box
-vagrant up
-vagrant ssh
+BOX_FILE="funtoo-x86_64-generic_64-stage3.box"
+BOX_NAME="funtoo-stage3"
+
+if [ -f "$BOX_FILE" ]
+then
+	echo "Suspending any running instances ..."
+	vagrant suspend
+	echo "Destroying current box ..."
+	vagrant destroy -f
+	echo "Removing '$BOX_NAME' ..."
+	vagrant box remove -f $BOX_NAME
+	echo "Adding '$BOX_NAME' ..."
+	vagrant box add --name $BOX_NAME
+	echo "Powerup '$BOX_NAME' ..." 
+	vagrant up
+	echo "Establishing SSH connection to '$BOX_NAME' ..."
+	vagrant ssh
+else
+	echo "There is no box file \"$BOX_FILE\" in the current directory. Please run './build.sh' to build the box."
+	exit 1
+fi
